Support `oldValue` instead of `updated` property
Change 36272952 prematurely added support to an unmerged
gerrit stream events change[1]. That gerrit change started
out with an `updated` property to indicate that an approval
was updated however thru reviews that change ended up adding
an `oldValue` property instead.  This property is used to
show the previous approval value when the score changed. Also
the property is only appended to the stream event if the
approval was changed so we can use this info to determine
whether the approval was updated.

[1] https://gerrit-review.googlesource.com/#/c/71051/